requirements.txt
uvicorn~=0.15.0
fastapi~=0.70.0
pydantic~=1.8.2
python-keycloak
PyYAML~=6.0
pandas~=1.3.4
pyodbc~=4.0.32
hdbcli~=2.11.14
python-keycloak~=0.26.1